+++ client/web/src/insights/hooks/use-insight/use-insight.ts
@@ -77,7 +77,7 @@ interface InsightsInputs {
     ownerId: string
 }
 
-export function createInsightFromSettings(input: InsightsInputs): Insight {
+export function createExtensionInsightFromSettings(input: InsightsInputs): Insight {
     const { insightKey, ownerId, insightConfiguration } = input
 
     return {