+++ cmd/repo-updater/repoupdater/server.go
@@ -15,7 +15,6 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/errcode"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/awscodecommit"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketserver"
@@ -38,9 +37,6 @@ type Server struct {
 	}
 	GitLabDotComSource interface {
 		GetRepo(ctx context.Context, projectWithNamespace string) (*types.Repo, error)
-	}
-	JVMPackagesSource interface {
-		GetRepo(ctx context.Context, artifactName string) (*types.Repo, error)
 	}
 	Scheduler interface {
 		UpdateOnce(id api.RepoID, name api.RepoName)
@@ -433,17 +429,6 @@ func (s *Server) remoteRepoSync(ctx context.Context, codehost *extsvc.CodeHost, 
 			if isUnauthorized(err) {
 				return &protocol.RepoLookupResult{
 					ErrorUnauthorized: true,
-				}, nil
-			}
-			return nil, err
-		}
-	case extsvc.JVMPackages:
-		artifactPath := strings.TrimPrefix(remoteName, "maven/")
-		repo, err = s.JVMPackagesSource.GetRepo(ctx, artifactPath)
-		if err != nil {
-			if errcode.IsNotFound(err) {
-				return &protocol.RepoLookupResult{
-					ErrorNotFound: true,
 				}, nil
 			}
 			return nil, err