+++ enterprise/cmd/precise-code-intel-worker/internal/worker/iface.go
@@ -56,6 +56,7 @@ type LSIFStore interface {
 	WriteReferences(ctx context.Context, bundleID int, monikerLocations chan semantic.MonikerLocations) error
 	WriteDocumentationPages(ctx context.Context, bundleID int, documentation chan *semantic.DocumentationPageData) error
 	WriteDocumentationPathInfo(ctx context.Context, bundleID int, documentation chan *semantic.DocumentationPathInfoData) error
+	WriteDocumentationMappings(ctx context.Context, bundleID int, mappings chan semantic.DocumentationMapping) error
 }
 
 type LSIFStoreShim struct {