+++ enterprise/internal/batches/resolvers/resolver.go
@@ -720,10 +720,44 @@ func (r *Resolver) BatchChanges(ctx context.Context, args *graphqlbackend.ListBa
 		}
 	}
 
+	if args.Repo != nil {
+		repoID, err := graphqlbackend.UnmarshalRepositoryID(*args.Repo)
+		if err != nil {
+			return nil, err
+		}
+		opts.RepoID = repoID
+	}
+
 	return &batchChangesConnectionResolver{
 		store: r.store,
 		opts:  opts,
 	}, nil
+}
+
+func (r *Resolver) RepoChangesetsStats(ctx context.Context, repo *graphql.ID) (graphqlbackend.RepoChangesetsStatsResolver, error) {
+	repoID, err := graphqlbackend.UnmarshalRepositoryID(*repo)
+	if err != nil {
+		return nil, err
+	}
+
+	stats, err := r.store.GetRepoChangesetsStats(ctx, repoID)
+	if err != nil {
+		return nil, err
+	}
+	return &repoChangesetsStatsResolver{stats: *stats}, nil
+}
+
+func (r *Resolver) RepoDiffStat(ctx context.Context, repo *graphql.ID) (*graphqlbackend.DiffStat, error) {
+	repoID, err := graphqlbackend.UnmarshalRepositoryID(*repo)
+	if err != nil {
+		return nil, err
+	}
+
+	diffStat, err := r.store.GetRepoDiffStat(ctx, repoID)
+	if err != nil {
+		return nil, err
+	}
+	return graphqlbackend.NewDiffStat(*diffStat), nil
 }
 
 func (r *Resolver) BatchChangesCodeHosts(ctx context.Context, args *graphqlbackend.ListBatchChangesCodeHostsArgs) (graphqlbackend.BatchChangesCodeHostConnectionResolver, error) {
@@ -878,6 +912,13 @@ func listChangesetOptsFromArgs(args *graphqlbackend.ListChangesetsArgs, batchCha
 	}
 	if args.OnlyArchived {
 		opts.OnlyArchived = args.OnlyArchived
+	}
+	if args.Repo != nil {
+		repoID, err := graphqlbackend.UnmarshalRepositoryID(*args.Repo)
+		if err != nil {
+			return opts, false, errors.Wrap(err, "unmarshalling repo id")
+		}
+		opts.RepoID = repoID
 	}
 
 	return opts, safe, nil