+++ internal/extsvc/codehost.go
@@ -21,13 +21,9 @@ var (
 	GitLabDotComURL = mustParseURL("https://gitlab.com")
 	GitLabDotCom    = NewCodeHost(GitLabDotComURL, TypeGitLab)
 
-	MavenURL    = &url.URL{Host: "maven"}
-	JVMPackages = NewCodeHost(MavenURL, TypeJVMPackages)
-
 	PublicCodeHosts = []*CodeHost{
 		GitHubDotCom,
 		GitLabDotCom,
-		JVMPackages,
 	}
 )